requirements.txt
streamlit==1.40.0
langchain==0.2.16
langchain-google-genai==1.0.7
python-dotenv==1.0.1
requests==2.32.3
